Revert "Remove the gpk-install-package-name binary"
This reverts commit bdeaf7f35511fa85d444deb9981d68e015acdea6.

This was part of the motivation for creating this fork. This small application is essential for custom distributions. For example, allow add an "Install Accessibility
Tools" desktop icon on menues, avoiding installing orca and many other packages to all users, but making it very easy to install it for those who need it